requirements.txt
Django>=2.0
psycopg2-binary
requests_oauthlib
gunicorn
whitenoise
dj-database-url
django-bootstrap4
user-agents
social-auth-app-django
beautifulsoup4
html5lib